backend/requirements.txt
# Core Django and Web Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Jac Programming Language Integration
jaclang>=0.2.0
jaseci>=1.4.0

# Database and Caching
psycopg2-binary>=2.9.0
redis>=4.5.0
django-redis>=5.2.0

# Celery for Background Tasks
celery[redis]>=5.2.0

# Environment and Configuration
python-decouple>=3.8
dj-database-url>=1.2.0

# Media and File Handling
Pillow>=9.5.0
django-storages>=1.13.0

# Development and Testing
django-debug-toolbar>=4.1.0
django-extensions>=3.2.0
pytest>=7.3.0
pytest-django>=4.5.0
factory-boy>=3.2.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0

# Security
django-security>=0.21.0

# API Documentation
drf-spectacular>=0.26.0